backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic[email]==2.8.2
pydantic-settings==2.4.0

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# Google Gemini (new unified SDK)
google-genai>=1.0.0

# Vector DB
qdrant-client==1.11.3

# PDF processing (pure Python, no system deps needed)
pypdf==4.3.1
PyMuPDF==1.24.10

# Async
httpx==0.27.2
aiofiles==24.1.0

# Utilities
python-dotenv==1.0.1
tenacity==9.0.0
